+++ tests/test_html_text.py
@@ -1,23 +1,17 @@
 # -*- coding: utf-8 -*-
 import pytest
-import lxml
 import glob
 
 from html_text import (extract_text, parse_html, cleaned_selector,
                        selector_to_text, NEWLINE_TAGS, DOUBLE_NEWLINE_TAGS)
 
 
-@pytest.fixture(params=[{
-    'guess_punct_space': True
-}, {
-    'guess_punct_space': False
-}, {
-    'guess_punct_space': True,
-    'guess_page_layout': True
-}, {
-    'guess_punct_space': False,
-    'guess_page_layout': True
-}])
+@pytest.fixture(params=[
+    {'guess_punct_space': True},
+    {'guess_punct_space': False},
+    {'guess_punct_space': True, 'guess_layout': True},
+    {'guess_punct_space': False, 'guess_layout': True}
+])
 def all_options(request):
     return request.param
 
@@ -67,8 +61,8 @@ def test_punct_whitespace():
 def test_punct_whitespace_preserved():
     html = (u'<div><span>по</span><span>ле</span>, and  ,  '
             u'<span>more </span>!<span>now</div>a (<b>boo</b>)')
-    assert (extract_text(
-        html, guess_punct_space=True) == u'по ле, and , more ! now a (boo)')
+    text = extract_text(html, guess_punct_space=True)
+    assert text == u'по ле, and , more ! now a (boo)'
 
 
 def test_bad_punct_whitespace():
@@ -98,7 +92,7 @@ def test_selector(all_options):
     assert selector_to_text(subsel, **all_options) == ''
 
 
-def test_guess_page_layout():
+def test_guess_layout():
     html = (u'<title>  title  </title><div>text_1.<p>text_2 text_3</p>'
             '<p id="demo"></p><ul><li>text_4</li><li>text_5</li></ul>'
             '<p>text_6<em>text_7</em>text_8</p>text_9</div>'
@@ -108,7 +102,7 @@ def test_guess_page_layout():
         'title text_1. text_2 text_3 text_4 text_5'
         ' text_6 text_7 text_8 text_9 ...text_10'))
     assert (extract_text(
-        html, guess_punct_space=False, guess_page_layout=True) == (
+        html, guess_punct_space=False, guess_layout=True) == (
             'title\n\n text_1.\n\n text_2 text_3\n\n text_4\n text_5'
             '\n\n text_6 text_7 text_8\n\n text_9\n\n ...text_10'))
     assert (extract_text(
@@ -116,33 +110,27 @@ def test_guess_page_layout():
         guess_punct_space=True) == ('title text_1. text_2 text_3 text_4 text_5'
                                     ' text_6 text_7 text_8 text_9...text_10'))
     assert (extract_text(
-        html, guess_punct_space=True, guess_page_layout=True) == (
+        html, guess_punct_space=True, guess_layout=True) == (
             'title\n\ntext_1.\n\ntext_2 text_3\n\ntext_4\ntext_5'
             '\n\ntext_6 text_7 text_8\n\ntext_9\n\n...text_10'))
 
 
 def test_adjust_newline():
     html = u'<div>text 1</div><p><div>text 2</div></p>'
-    assert (extract_text(html, guess_punct_space=True,
-                         guess_page_layout=True) == ('text 1\n\ntext 2'))
+    assert extract_text(html, guess_layout=True) == 'text 1\n\ntext 2'
 
 
 def test_personalize_newlines_sets():
     html = (u'<span><span>text<a>more</a>'
             '</span>and more text <a> and some more</a> <a></a> </span>')
-    assert (extract_text(
-        html,
-        guess_punct_space=True,
-        guess_page_layout=True,
-        newline_tags=NEWLINE_TAGS | {'a'}
-        ) == 'text\nmore\nand more text\nand some more')
 
-    assert (extract_text(
-        html,
-        guess_punct_space=True,
-        guess_page_layout=True,
-        double_newline_tags=DOUBLE_NEWLINE_TAGS | {'a'}
-        ) == 'text\n\nmore\n\nand more text\n\nand some more')
+    text = extract_text(html, guess_layout=True,
+                        newline_tags=NEWLINE_TAGS | {'a'})
+    assert text == 'text\nmore\nand more text\nand some more'
+
+    text = extract_text(html, guess_layout=True,
+                        double_newline_tags=DOUBLE_NEWLINE_TAGS | {'a'})
+    assert text == 'text\n\nmore\n\nand more text\n\nand some more'
 
 
 def test_webpages():
@@ -153,5 +141,4 @@ def test_webpages():
             html = f_in.read()
         with open(extr, 'r', encoding='utf8') as f_in:
             expected = f_in.read()
-        assert (extract_text(
-            html, guess_punct_space=True, guess_page_layout=True) == expected)
+        assert extract_text(html, guess_layout=True) == expected